registry path for scoped package tarball should include scope subdirectory under dash directory
The JSON sent by `npm publish` contains the property `versions.XXX.dist.tarball` which contains the scope name twice for a scoped package; for example: "http://localhost:15443/@chad/myhowtonpm/-/@chad/myhowtonpm-1.0.0.tgz". The mock registry server attempts to remove the second occurrence of `@scope` from the file path, which causes an error if users try to create another package into which they install the package created in the workshopper using the mock server. The path should be left as-is to match the tarball property in `body.json` so npm can GET it. Removing this code has the additional benefit of fixing a bug on Windows because the regex does not consider backslashes in the path.

See: https://github.com/workshopper/how-to-npm/pull/68
Fixes: https://github.com/workshopper/how-to-npm/issues/122
Fixes: https://github.com/nodeschool/discussions/issues/1561
Fixes: https://github.com/nodeschool/discussions/issues/2049